Restore equipment getter return width
Model pokemonGetSoubiItemDataId as a full-width return so the two GBA summary paths perform their target-visible narrowing after preserving the call result.\n\nConstraint: Call sites continue to treat the semantic item identifier as 16-bit where required.\nRejected: Force register moves with volatile temporaries | would encode compiler artifacts instead of the recovered ABI\nConfidence: high\nScope-risk: narrow\nDirective: Keep explicit use-site narrowing rather than narrowing the function declaration.\nTested: ninja build/GC6E01/src/game/gbaCommunication_candidate_800965C8_gc20.o; ninja -j4 build/GC6E01/report.json; fn_800965C8 fuzzy 97.62981% -> 97.88221%\nNot-tested: full link/runtime

+++ src/game/gbaCommunication.c
@@ -88,7 +88,7 @@ extern void* menuSpriteBiosGetPtr(u16 id);
 extern u16* windowGetKeyInfo(void);
 extern void* pokemonDataBiosGetPtr(u32 id);
 extern u8 pokemonBiosGetCatchBallId(void* pokemon);
-extern u16 pokemonGetSoubiItemDataId(void* pokemon);
+extern u32 pokemonGetSoubiItemDataId(void* pokemon);
 extern u32 pokemonDataBiosGetName(void* bios);
 extern u32 GSmsgGetGSchar(u32 id);
 extern u32 GSmsgGetRect(u32 id);
